fix(zuuli): keep recovery copy neutral
