+++ wallet/zuuli/package.json
@@ -9,7 +9,7 @@
     "dev": "npm run wasm:build && vite",
     "build": "npm run wasm:build && tsc -p tsconfig.build.json && vite build && npm run wasm:verify-dist",
     "preview": "vite preview",
-    "test": "vitest run && node --test scripts/safe-area-contract.node-test.mjs scripts/android-device-catalog.node-test.mjs scripts/android-release-artifact.node-test.mjs scripts/aab-payload-digest.node-test.mjs scripts/auth-session-boundary.node-test.mjs scripts/mermaid-security.node-test.mjs scripts/send-review-boundary.node-test.mjs scripts/mobile-webview-authority.node-test.mjs scripts/seed-capture-boundary.node-test.mjs scripts/ui-copy-truncation.node-test.mjs scripts/ui-copy-source-audit.node-test.mjs scripts/fixture-privacy.node-test.mjs scripts/apple-credential-boundary.node-test.mjs scripts/macos-keychain-entitlements.node-test.mjs scripts/artifact-sbom.node-test.mjs scripts/release-tag-identity.node-test.mjs scripts/status-freshness.node-test.mjs scripts/wasm-boundary.node-test.mjs scripts/messaging-contract.node-test.mjs && node scripts/apple-credential-boundary.mjs && node scripts/macos-keychain-entitlements.mjs && playwright test",
+    "test": "vitest run && node --test scripts/safe-area-contract.node-test.mjs scripts/android-device-catalog.node-test.mjs scripts/android-release-artifact.node-test.mjs scripts/aab-payload-digest.node-test.mjs scripts/auth-session-boundary.node-test.mjs scripts/mermaid-security.node-test.mjs scripts/send-review-boundary.node-test.mjs scripts/mobile-webview-authority.node-test.mjs scripts/seed-capture-boundary.node-test.mjs scripts/ui-copy-truncation.node-test.mjs scripts/fixture-privacy.node-test.mjs scripts/apple-credential-boundary.node-test.mjs scripts/macos-keychain-entitlements.node-test.mjs scripts/artifact-sbom.node-test.mjs scripts/release-tag-identity.node-test.mjs scripts/status-freshness.node-test.mjs scripts/wasm-boundary.node-test.mjs scripts/messaging-contract.node-test.mjs && node scripts/apple-credential-boundary.mjs && node scripts/macos-keychain-entitlements.mjs && playwright test",
     "test:seed-capture": "vitest run src/lib/wallet/sensitive-seed.test.ts src/lib/wallet/sensitive-entry.test.ts src/lib/wallet/sensitive-entry-bridges.test.ts src/lib/wallet/sensitive-entry-hooks.test.tsx src/lib/wallet/zuuallet-created-seed.test.ts src/features/wallet/Onboarding.sensitive-entry.test.tsx && npm --prefix ../zuuallet run test:sensitive-entry && node --test scripts/seed-capture-boundary.node-test.mjs && node scripts/seed-capture-boundary.mjs",
     "test:legacy-import-preview": "vitest run src/lib/wallet/legacy-import-preview-bridge.test.ts src/components/layout/LegacyWalletNotice.test.tsx",
     "test:legacy-import-preview:browser": "playwright test tests/legacy-import-preview.pw.ts",
@@ -19,7 +19,7 @@
     "test:apple-credential-boundary": "node --test scripts/apple-credential-boundary.node-test.mjs && node scripts/apple-credential-boundary.mjs",
     "test:artifact-sbom": "node --test scripts/artifact-sbom.node-test.mjs",
     "test:viewport": "playwright test tests/viewport.pw.ts",
-    "test:ui-copy": "node --test scripts/ui-copy-truncation.node-test.mjs scripts/ui-copy-source-audit.node-test.mjs && playwright test tests/viewport.pw.ts tests/ui-copy.pw.ts",
+    "test:ui-copy": "node --test scripts/ui-copy-truncation.node-test.mjs && playwright test tests/viewport.pw.ts tests/ui-copy.pw.ts",
     "test:asc-testflight": "node --test scripts/asc-testflight.node-test.mjs",
     "test:store-listing": "node --test scripts/store-contract.node-test.mjs scripts/store-state-audit.node-test.mjs",
     "test:status-freshness": "node --test scripts/status-freshness.node-test.mjs",
